add thread safety assertions to `getLuaState`
split `CoreDefs.h` from `Core.h` to keep `Core` from being loaded into `RemoteClient`
