+++ library/include/Core.h
@@ -25,6 +25,7 @@ distribution.
 #pragma once
 
 #include "Console.h"
+#include "CoreDefs.h"
 #include "Export.h"
 #include "Hooks.h"
 
@@ -75,31 +76,6 @@ namespace DFHack
     {
         struct Hide;
     }
-
-    enum command_result
-    {
-        CR_LINK_FAILURE = -3,    // RPC call failed due to I/O or protocol error
-        CR_NEEDS_CONSOLE = -2,   // Attempt to call interactive command without console
-        CR_NOT_IMPLEMENTED = -1, // Command not implemented, or plugin not loaded
-        CR_OK = 0,               // Success
-        CR_FAILURE = 1,          // Failure
-        CR_WRONG_USAGE = 2,      // Wrong arguments or ui state
-        CR_NOT_FOUND = 3         // Target object not found (for RPC mainly)
-    };
-
-    enum state_change_event
-    {
-        SC_UNKNOWN = -1,
-        SC_WORLD_LOADED = 0,
-        SC_WORLD_UNLOADED = 1,
-        SC_MAP_LOADED = 2,
-        SC_MAP_UNLOADED = 3,
-        SC_VIEWSCREEN_CHANGED = 4,
-        SC_CORE_INITIALIZED = 5,
-        SC_BEGIN_UNLOAD = 6,
-        SC_PAUSED = 7,
-        SC_UNPAUSED = 8
-    };
 
     class DFHACK_EXPORT PerfCounters
     {
@@ -244,7 +220,10 @@ namespace DFHack
 
         PerfCounters perf_counters;
 
-        lua_State* getLuaState() { return State; }
+        lua_State* getLuaState(bool bypass_assertion = false) {
+            assert(bypass_assertion || isSuspended());
+            return State;
+        }
 
     private:
         DFHack::Console con;
@@ -400,9 +379,11 @@ namespace DFHack
             if (!owns_lock())
                 return;
             /* Restore core owner to previous value */
-            core.ownerThread.store(tid, std::memory_order_release);
             if (tid == std::thread::id{})
                 Lua::Core::Reset(core.getConsole(), "suspend");
+            core.ownerThread.store(tid, std::memory_order_release);
+            //if (tid == std::thread::id{})
+            //    Lua::Core::Reset(core.getConsole(), "suspend");
             parent_t::unlock();
         }
