+++ library/include/RemoteClient.h
@@ -25,7 +25,7 @@ distribution.
 #pragma once
 #include "Export.h"
 #include "ColorText.h"
-#include "Core.h"
+#include "CoreDefs.h"
 
 class CPassiveSocket;
 class CActiveSocket;
